Revert "Separate the ubus module from the main package"
