+++ first-boot-wizard-daemon/files/usr/share/rpcd/acl.d/firstbootwizard.json
@@ -1,0 +1,15 @@
+{
+    "firstbootwizard": {
+        "description": "firstbootwizard ubus interface auth",
+        "read": {
+            "ubus": {
+                    "firstbootwizard": [ "*" ]
+            }
+        },
+        "write": {
+            "ubus": {
+                    "firstbootwizard": [ "*" ]
+            }
+        }
+    }
+}
