+++ first-boot-wizard-daemon/files/var/lock/first_run
@@ -1,0 +1,1 @@
+true
