+++ first-boot-wizard-gui/Makefile
@@ -1,11 +1,10 @@
 include $(TOPDIR)/rules.mk
+include $(INCLUDE_DIR)/package.mk
 
-PKG_NAME:=first-boot-wizard
+PKG_NAME:=first-boot-wizard-gui
 PKG_VERSION=$(GIT_COMMIT_DATE)-$(GIT_COMMIT_TSTAMP)
 GIT_COMMIT_DATE:=$(shell git log -n 1 --pretty=%ad --date=short . )
 GIT_COMMIT_TSTAMP:=$(shell git log -n 1 --pretty=%at . )
-
-include $(INCLUDE_DIR)/package.mk
 
 define Package/$(PKG_NAME)
     TITLE:=$(PKG_NAME)
@@ -16,7 +15,7 @@ define Package/$(PKG_NAME)
 endef
 
 define Package/$(PKG_NAME)/description
-    Scans surrounding LibreMesh networks and joins them.
+    Allows the user to initialize a new network.
 endef
 
 define Build/Compile
