+++ first-boot-wizard-gui/files/www/start/index.html
@@ -1,0 +1,11 @@
+<html>
+  <head>
+    <title>First Boot Wizard</title>
+  </head>
+  <body>
+    <form method="get" action="first_boot_wizard/create_network_profile">
+      <label>Network's name:<input name="name"></label>
+      <submit />
+    </form>
+  </body>
+</html>
