--- first-boot-wizard/files/bin/firstbootwizard
@@ -1,6 +1,0 @@
-#!/usr/bin/lua
-
-local fbw = require('firstbootwizard')
-
-print("[FBW] start.")
-get_all_networks()
